+++ src/main/java/com/isfive/usearth/domain/board/service/PostService.java
@@ -11,7 +11,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.isfive.usearth.domain.board.annotation.Retry;
+import com.isfive.usearth.annotation.Retry;
 import com.isfive.usearth.domain.board.dto.PostCommentResponse;
 import com.isfive.usearth.domain.board.dto.PostResponse;
 import com.isfive.usearth.domain.board.dto.PostsResponse;
